move ol functions into openlayersviewer
