+++ package.json
@@ -59,6 +59,7 @@
     "jest": "^30.3.0",
     "jest-environment-jsdom": "^30.4.1",
     "lerna": "^9.0.7",
+    "maplibre-gl": "^6.4.1",
     "npm-run-all": "^4.1.5",
     "prettier": "~3.8.0",
     "py2vega-ts": "^0.1.2",
