+++ packages/base/src/mainview/mainView.tsx
@@ -4,11 +4,6 @@ import {
   IAnnotation,
   IAnnotationModel,
   IDict,
-  IGeoTiffSource,
-  IGeoZarrSource,
-  IHillshadeLayer,
-  IImageLayer,
-  IImageSource,
   IJGISFilterItem,
   IJGISFormSchemaRegistry,
   IJGISLayer,
@@ -25,30 +20,13 @@ import {
   IJupyterGISDoc,
   IJupyterGISDocChange,
   IJupyterGISModel,
-  IRasterDemSource,
-  IRasterLayer,
-  IRasterSource,
-  IShapefileSource,
-  IStacLayer,
   IVectorLayer,
-  IVectorTileLayer,
-  IVectorTileSource,
-  IGeoParquetSource,
-  IGeoTiffLayer,
-  IGeoZarrLayer,
   JgisCoordinates,
   JupyterGISModel,
   IMarkerSource,
-  IStorySegmentLayer,
-  IWmsTileSource,
   IJupyterGISSettings,
   DEFAULT_PROJECTION,
-  IViewState,
-  IOpenEOTileSource,
-  IOpenEOTileLayer,
-  IGrammarSymbologyState,
 } from '@jupytergis/schema';
-import { showErrorMessage } from '@jupyterlab/apputils';
 import type { ILoggerRegistry } from '@jupyterlab/logconsole';
 import { IObservableMap, ObservableMap } from '@jupyterlab/observables';
 import { User } from '@jupyterlab/services';
@@ -56,23 +34,14 @@ import { IStateDB } from '@jupyterlab/statedb';
 import { CommandRegistry } from '@lumino/commands';
 import { JSONValue, UUID } from '@lumino/coreutils';
 import { ContextMenu, Menu } from '@lumino/widgets';
-import {
-  Collection,
-  Geolocation,
-  MapBrowserEvent,
-  Map as OlMap,
-  VectorTile,
-  View,
-  getUid,
-} from 'ol';
+import { Geolocation, MapBrowserEvent, Map as OlMap, View } from 'ol';
 import Feature, { FeatureLike } from 'ol/Feature';
 import type { GeolocationError } from 'ol/Geolocation';
-import TileState from 'ol/TileState';
 import { FullScreen, ScaleLine, Zoom, Control, Rotate } from 'ol/control';
 import { Coordinate } from 'ol/coordinate';
 import { singleClick } from 'ol/events/condition';
-import { getCenter, getSize } from 'ol/extent';
-import { GeoJSON, MVT } from 'ol/format';
+import { getCenter } from 'ol/extent';
+import { GeoJSON } from 'ol/format';
 import { Geometry, Point } from 'ol/geom';
 import { Type } from 'ol/geom/Geometry';
 import {
@@ -93,44 +62,21 @@ import type Interaction from 'ol/interaction/Interaction';
 import Modify from 'ol/interaction/Modify';
 import Snap from 'ol/interaction/Snap';
 import {
-  Image as ImageLayer,
-  Layer,
   Vector as VectorLayer,
   VectorImage as VectorImageLayer,
-  VectorTile as VectorTileLayer,
   WebGLTile as RasterLayer,
 } from 'ol/layer';
 import LayerGroup from 'ol/layer/Group';
-import TileLayer from 'ol/layer/Tile';
 import {
   fromLonLat,
   get as getProjection,
   toLonLat,
   transformExtent,
 } from 'ol/proj';
-import { register } from 'ol/proj/proj4.js';
 import RenderFeature, { toGeometry } from 'ol/render/Feature';
-import {
-  GeoTIFF as GeoTIFFSource,
-  ImageTile as ImageTileSource,
-  Source,
-  TileWMS as TileWMSSource,
-  Vector as VectorSource,
-  VectorTile as VectorTileSource,
-  XYZ as XYZSource,
-  Tile as TileSource,
-} from 'ol/source';
-import GeoZarr from 'ol/source/GeoZarr';
-import Static from 'ol/source/ImageStatic';
-import { TileSourceEvent } from 'ol/source/Tile';
+import { Vector as VectorSource } from 'ol/source';
 import { Fill, Icon, Stroke, Style } from 'ol/style';
 import CircleStyle from 'ol/style/Circle';
-import { Rule } from 'ol/style/flat';
-//@ts-expect-error no types for ol-pmtiles
-import { PMTilesRasterSource, PMTilesVectorSource } from 'ol-pmtiles';
-import StacLayer from 'ol-stac';
-import proj4 from 'proj4';
-import proj4list from 'proj4-list';
 import * as React from 'react';
 
 import { CommandIDs } from '@/src/constants';
@@ -143,15 +89,7 @@ import {
   isVerticalScrollPresentation,
 } from '@/src/features/story/presentation/getStoryPresentationMode';
 import { useIsMobile } from '@/src/shared/hooks/useIsMobile';
-import { markerIcon } from '@/src/shared/icons';
-import {
-  debounce,
-  INTERNAL_PROXY_BASE,
-  isJupyterLite,
-  isLightTheme,
-  loadFile,
-  throttle,
-} from '@/src/tools';
+import { debounce, isLightTheme, throttle } from '@/src/tools';
 import StatusBar from '@/src/workspace/statusbar/StatusBar';
 import { ClientPointer } from './CollaboratorPointers';
 import TemporalSlider from './TemporalSlider';
@@ -168,35 +106,9 @@ import { MainViewModel } from './mainviewmodel';
 import { createMapViewer, IMapViewer, MapViewerType } from './mapviewer';
 import { ensureHighlightLayer } from '../features/identify/utils/highlightLayer';
 import { buildHighlightStyle } from '../features/identify/utils/highlightStyle';
-import {
-  OpenEOTileLayer,
-  OpenEOTileSource,
-  openEOEvents,
-} from '../features/layers/openeo/OpenEOTileLayer';
-import { grammarToOLLayer } from '../features/layers/symbology/grammarToOLLayer';
-import {
-  extractEncodingFieldValues,
-  grammarToOLStyle,
-} from '../features/layers/symbology/grammarToOLStyle';
-import { DEFAULT_FLAT_STYLE } from '../features/layers/symbology/styleBuilder';
-import {
-  buildZarrColorStyle,
-  getBandInfoFromZarr,
-  getDefaultRGBBands,
-  IZarrBandInfo,
-} from '../features/layers/symbology/zarrBandDiscovery';
+import { openEOEvents } from '../features/layers/openeo/OpenEOTileLayer';
 import type { IStoryViewerPanelHandle } from '../features/story/StoryViewerPanel';
 import type { IListStorySegmentTransition } from '../features/story/types/types';
-
-type OlLayerTypes =
-  | TileLayer
-  | VectorLayer
-  | VectorImageLayer
-  | VectorTileLayer
-  | RasterLayer
-  | StacLayer
-  | ImageLayer<any>
-  | LayerGroup;
 
 const drawInteractionStyle = new Style({
   fill: new Fill({
@@ -334,7 +246,6 @@ export class MainView extends React.Component<IMainViewProps, IStates> {
       this._handleIdentifiedFeaturesChanged,
       this,
     );
-    this._model.zoomToPositionSignal.connect(this._onZoomToPosition, this);
     this._model.settingsChanged.connect(this._onSettingsChanged, this);
     this._model.updateLayerSignal.connect(this._triggerLayerUpdate, this);
     this._model.addFeatureAsMsSignal.connect(this._convertFeatureToMs, this);
@@ -392,8 +303,6 @@ export class MainView extends React.Component<IMainViewProps, IStates> {
       segmentTransition: null,
     };
 
-    this._sources = [];
-    this._loadingLayers = new Set();
     this._commands = new CommandRegistry();
     this._contextMenu = new ContextMenu({
       commands: this._commands,
@@ -421,6 +330,11 @@ export class MainView extends React.Component<IMainViewProps, IStates> {
     const zoom = options.zoom !== undefined ? options.zoom : 1;
 
     await this.generateMap(center, zoom, projection);
+    this._model.zoomToPositionSignal.connect(
+      this._mapViewer.onZoomToPosition,
+      this._mapViewer,
+    );
+
     this._handleRemoteUserChanged();
     this._handlePointerChanged();
     this._handleTemporalControllerActiveChanged();
@@ -506,7 +420,6 @@ export class MainView extends React.Component<IMainViewProps, IStates> {
     this._stopLocationIndicator();
     if (this._mapViewer) {
       this._mapViewer.destroy();
-      this._mapViewer = null;
     }
 
     this._mainViewModel.dispose();
@@ -547,11 +460,11 @@ export class MainView extends React.Component<IMainViewProps, IStates> {
       this._lastMapViewerType = mapViewerType;
 
       // Create the map viewer using abstraction
-      this._mapViewer = await createMapViewer(mapViewerType);
+      this._mapViewer = await createMapViewer(mapViewerType, this._model);
 
       await this._mapViewer.initialize(this.divRef.current, {
         projection,
-        center,
+        center: [center[0], center[1]],
         zoom,
         rotation: 0,
       });
@@ -589,7 +502,7 @@ export class MainView extends React.Component<IMainViewProps, IStates> {
 
         const layerId = UUID.uuid4();
 
-        this.addSource(sourceId, sourceModel);
+        this._mapViewer.addSource(sourceId, sourceModel);
 
         this._model.sharedModel.addSource(sourceId, sourceModel);
 
@@ -605,7 +518,11 @@ export class MainView extends React.Component<IMainViewProps, IStates> {
           },
         };
 
-        this.addLayer(layerId, layerModel, this.getLayerIDs().length);
+        this._mapViewer.addLayer(
+          layerId,
+          layerModel,
+          this._mapViewer.getLayerIDs().length,
+        );
         this._model.addLayer(layerId, layerModel);
       });
 
@@ -719,7 +636,7 @@ export class MainView extends React.Component<IMainViewProps, IStates> {
         .addEventListener('pointermove', this._onPointerMove.bind(this));
 
       if (JupyterGISModel.getOrderedLayerIds(this._model).length !== 0) {
-        await this._updateLayersImpl(
+        await this._mapViewer.updateLayersImpl(
           JupyterGISModel.getOrderedLayerIds(this._model),
         );
         const options = this._model.getOptions();
@@ -857,7 +774,7 @@ export class MainView extends React.Component<IMainViewProps, IStates> {
           return false;
         }
         const selectedLayerId = Object.keys(selectedLayers)[0];
-        const expected = this.getLayer(selectedLayerId);
+        const expected = this._mapViewer.getLayer(selectedLayerId);
         if (layer === expected) {
           return true;
         }
@@ -887,7 +804,7 @@ export class MainView extends React.Component<IMainViewProps, IStates> {
         ? Object.keys(selectedLayers)[0]
         : undefined;
       const mapLayer = selectedLayerId
-        ? this.getLayer(selectedLayerId)
+        ? this._mapViewer.getLayer(selectedLayerId)
         : undefined;
 
       // For LayerGroup (multi-layer grammar), collect style functions from
@@ -1049,1276 +966,6 @@ export class MainView extends React.Component<IMainViewProps, IStates> {
     });
   };
 
-  /**
-   * Add a source in the map.
-   *
-   * @param id - the source id.
-   * @param source - the source object.
-   */
-  async addSource(id: string, source: IJGISSource): Promise<void> {
-    this._log('info', `Loading source "${source.name ?? id}" (${source.type})`);
-    let newSource;
-
-    try {
-      switch (source.type) {
-        case 'RasterSource': {
-          const sourceParameters = source.parameters as IRasterSource;
-
-          const pmTiles =
-            sourceParameters.url.endsWith('.pmtiles') ||
-            sourceParameters.url.endsWith('pmtiles.gz');
-          const url = this.computeSourceUrl(source);
-
-          if (!pmTiles) {
-            newSource = new XYZSource({
-              interpolate: sourceParameters.interpolate,
-              attributions: sourceParameters.attribution,
-              minZoom: sourceParameters.minZoom,
-              maxZoom: sourceParameters.maxZoom,
-              tileSize: 256,
-              url: url,
-            });
-          } else {
-            newSource = new PMTilesRasterSource({
-              interpolate: sourceParameters.interpolate,
-              attributions: sourceParameters.attribution,
-              tileSize: 256,
-              url: url,
-            });
-          }
-
-          break;
-        }
-
-        case 'RasterDemSource': {
-          const sourceParameters = source.parameters as IRasterDemSource;
-
-          newSource = new ImageTileSource({
-            interpolate: sourceParameters.interpolate,
-            url: this.computeSourceUrl(source),
-            attributions: sourceParameters.attribution,
-          });
-
-          break;
-        }
-
-        case 'VectorTileSource': {
-          const sourceParameters = source.parameters as IVectorTileSource;
-
-          const pmTiles =
-            sourceParameters.url.endsWith('.pmtiles') ||
-            sourceParameters.url.endsWith('pmtiles.gz');
-          const url = this.computeSourceUrl(source);
-
-          if (!pmTiles) {
-            const vtSourceOptions: ConstructorParameters<
-              typeof VectorTileSource
-            >[0] = {
-              attributions: sourceParameters.attribution,
-              minZoom: sourceParameters.minZoom,
-              maxZoom: sourceParameters.maxZoom,
-              url: url,
-              format: new MVT({
-                featureClass: RenderFeature,
-              }),
-            };
-
-            if (sourceParameters.useProxy) {
-              const extraHeaders = sourceParameters.httpHeaders ?? {};
-              const headersParam =
-                Object.keys(extraHeaders).length > 0
-                  ? `&headers=${encodeURIComponent(JSON.stringify(extraHeaders))}`
-                  : '';
-
-              const proxyBase = isJupyterLite()
-                ? `${this._model.jgisSettings.proxyUrl}/`
-                : `${INTERNAL_PROXY_BASE}`;
-
-              vtSourceOptions.tileLoadFunction = (tile, tileUrl) => {
-                const vtTile = tile as VectorTile<RenderFeature>;
-                const proxyUrl = `${proxyBase}?url=${encodeURIComponent(tileUrl)}${headersParam}`;
-                vtTile.setLoader((extent, _resolution, projection) => {
-                  return fetch(proxyUrl)
-                    .then(response => {
-                      if (!response.ok) {
-                        throw new Error(
-                          `Tile proxy request failed: ${response.status} ${response.statusText}`,
-                        );
-                      }
-                      return response.arrayBuffer();
-                    })
-                    .then(data => {
-                      const features = vtTile.getFormat().readFeatures(data, {
-                        extent,
-                        featureProjection: projection,
-                      });
-                      vtTile.setFeatures(features);
-                      this._log('debug', `Proxy tile loaded: ${tileUrl}`);
-                      return features;
-                    })
-                    .catch((err: any) => {
-                      this._log(
-                        'error',
-                        `Proxy tile error for ${tileUrl}: ${err.message}`,
-                      );
-                      tile.setState(TileState.ERROR);
-                      return [];
-                    });
-                });
-              };
-            }
-
-            newSource = new VectorTileSource(vtSourceOptions);
-          } else {
-            newSource = new PMTilesVectorSource({
-              attributions: sourceParameters.attribution,
-              url: url,
-            });
-          }
-
-          newSource.on('tileloadend', (event: TileSourceEvent) => {
-            const tile = event.tile as VectorTile<FeatureLike>;
-            const features = tile.getFeatures();
-
-            if (features && features.length > 0) {
-              this._model.syncTileFeatures({
-                sourceId: id,
-                features,
-              });
-            }
-          });
-
-          break;
-        }
-
-        case 'OpenEOTileSource': {
-          const sourceParameters = source.parameters as IOpenEOTileSource;
-
-          newSource = new OpenEOTileSource({
-            serverUrl: sourceParameters.serverUrl ?? '',
-            authBearer: sourceParameters.authBearer,
-            processGraph: sourceParameters.processGraph,
-          });
-
-          break;
-        }
-
-        case 'GeoJSONSource': {
-          const data =
-            source.parameters?.data ||
-            (await loadFile({
-              filepath: source.parameters?.path,
-              type: 'GeoJSONSource',
-              model: this._model,
-            }));
-
-          const format = new GeoJSON({
-            featureProjection: this._Map.getView().getProjection(),
-          });
-
-          const featureArray = format.readFeatures(data, {
-            featureProjection: this._Map.getView().getProjection(),
-          });
-
-          const featureCollection = new Collection(featureArray);
-
-          featureCollection.forEach(feature => {
-            feature.setId(getUid(feature));
-          });
-
-          newSource = new VectorSource({
-            features: featureCollection,
-          });
-
-          break;
-        }
-
-        case 'ShapefileSource': {
-          const parameters = source.parameters as IShapefileSource;
-
-          const geojson = await loadFile({
-            filepath: parameters.path,
-            type: 'ShapefileSource',
-            model: this._model,
-          });
-
-          const geojsonData = Array.isArray(geojson) ? geojson[0] : geojson;
-
-          const format = new GeoJSON();
-
-          newSource = new VectorSource({
-            features: format.readFeatures(geojsonData, {
-              dataProjection: 'EPSG:4326',
-              featureProjection: this._Map.getView().getProjection(),
-            }),
-          });
-          break;
-        }
-
-        case 'ImageSource': {
-          const sourceParameters = source.parameters as IImageSource;
-
-          // Convert lon/lat array to extent
-          // Get lon/lat from source coordinates
-          const leftSide = Math.min(
-            ...sourceParameters.coordinates.map(corner => corner[0]),
-          );
-          const bottomSide = Math.min(
-            ...sourceParameters.coordinates.map(corner => corner[1]),
-          );
-          const rightSide = Math.max(
-            ...sourceParameters.coordinates.map(corner => corner[0]),
-          );
-          const topSide = Math.max(
-            ...sourceParameters.coordinates.map(corner => corner[1]),
-          );
-
-          // Convert lon/lat to OpenLayer coordinates
-          const topLeft = fromLonLat([leftSide, topSide]);
-          const bottomRight = fromLonLat([rightSide, bottomSide]);
-
-          // Get extent from coordinates
-          const minX = topLeft[0];
-          const maxY = topLeft[1];
-          const maxX = bottomRight[0];
-          const minY = bottomRight[1];
-
-          const extent = [minX, minY, maxX, maxY];
-
-          const imageUrl = await loadFile({
-            filepath: sourceParameters.path,
-            type: 'ImageSource',
-            model: this._model,
-          });
-
-          newSource = new Static({
-            interpolate: sourceParameters.interpolate,
-            imageExtent: extent,
-            url: imageUrl,
-            crossOrigin: '',
-          });
-
-          break;
-        }
-
-        case 'GeoTiffSource': {
-          const sourceParameters = source.parameters as IGeoTiffSource;
-
-          const addNoData = (url: (typeof sourceParameters.urls)[0]) => {
-            return { ...url, nodata: 0 };
-          };
-          const sources = await Promise.all(
-            sourceParameters.urls.map(async sourceInfo => {
-              const isRemote =
-                sourceInfo.url?.startsWith('http://') ||
-                sourceInfo.url?.startsWith('https://');
-              const isDataUrl = sourceInfo.url?.startsWith('data:');
-
-              if (isRemote) {
-                return {
-                  ...addNoData(sourceInfo),
-                  url: sourceInfo.url,
-                };
-              } else if (isDataUrl) {
-                // Inline base64 GeoTIFF embedded in the .jGIS doc.
-                const blob = await (await fetch(sourceInfo.url!)).blob();
-                return {
-                  ...addNoData(sourceInfo),
-                  url: URL.createObjectURL(blob),
-                };
-              } else {
-                const geotiff = await loadFile({
-                  filepath: sourceInfo.url ?? '',
-                  type: 'GeoTiffSource',
-                  model: this._model,
-                });
-                return {
-                  ...addNoData(sourceInfo),
-                  geotiff,
-                  url: URL.createObjectURL(geotiff.file),
-                };
-              }
-            }),
-          );
-
-          newSource = new GeoTIFFSource({
-            interpolate: sourceParameters.interpolate,
-            sources,
-            normalize: sourceParameters.normalize,
-            wrapX: sourceParameters.wrapX,
-            ...(sourceParameters.projection
-              ? { projection: sourceParameters.projection }
-              : {}),
-          });
-
-          break;
-        }
-
-        case 'GeoZarrSource': {
-          const sourceParameters = source.parameters as IGeoZarrSource;
-
-          let bands: string[] = sourceParameters.bands || [];
-
-          if (bands.length === 0) {
-            let bandInfo: IZarrBandInfo[] = [];
-
-            try {
-              bandInfo = await getBandInfoFromZarr(sourceParameters.url);
-            } catch (err) {
-              console.warn('Failed to auto-detect Zarr bands:', err);
-            }
-
-            if (bandInfo.length > 0) {
-              bands = getDefaultRGBBands(bandInfo);
-
-              // Persist to model for future loads
-              const updatedSource: IJGISSource = {
-                ...source,
-                parameters: {
-                  ...sourceParameters,
-                  bands,
-                },
-              };
-
-              this._model.sharedModel.updateSource(id, updatedSource);
-            } else {
-              console.warn('No bands detected from Zarr store');
-              bands = sourceParameters.bands?.length
-                ? sourceParameters.bands
-                : ['b04', 'b03', 'b02'];
-            }
-          }
-
-          newSource = new GeoZarr({
-            url: sourceParameters.url,
-            bands,
-            wrapX: sourceParameters.wrapX,
-          });
-
-          break;
-        }
-
-        case 'GeoPackageVectorSource': {
-          const sourceParameters = source.parameters;
-
-          if (!sourceParameters) {
-            throw new Error('GeoPackageSource has no parameters');
-          }
-
-          const tableMap = await loadFile({
-            filepath: sourceParameters.path,
-            type: 'GeoPackageVectorSource',
-            model: this._model,
-          });
-
-          const table = tableMap[sourceParameters.tables];
-          const vectorSource = table.source;
-          vectorSource['projection'] = getProjection(
-            sourceParameters.projection,
-          );
-          newSource = vectorSource;
-          break;
-        }
-
-        case 'GeoPackageRasterSource': {
-          const sourceParameters = source.parameters;
-
-          if (!sourceParameters) {
-            throw new Error('GeoPackageSource has no parameters');
-          }
-
-          const tableMap = await loadFile({
-            filepath: sourceParameters.path,
-            type: 'GeoPackageRasterSource',
-            model: this._model,
-          });
-
-          const { gpr, tileDao } = tableMap[sourceParameters.tables];
-
-          const rasterSource = new XYZSource({
-            minZoom: sourceParameters.minZoom ?? tileDao.minWebMapZoom,
-            maxZoom: sourceParameters.maxZoom ?? tileDao.maxWebMapZoom,
-            interpolate: sourceParameters.interpolate,
-            url: '{z},{x},{y}',
-            tileLoadFunction(tile: any, src) {
-              const [z, x, y] = src.split(',').map(Number);
-              gpr
-                .getTile(x, y, z)
-                .then((dataUri: any) => (tile.getImage().src = dataUri));
-            },
-            attributions: sourceParameters.attribution,
-          });
-
-          newSource = rasterSource;
-          break;
-        }
-
-        case 'GeoParquetSource': {
-          const parameters = source.parameters as IGeoParquetSource;
-
-          const geojson = await loadFile({
-            filepath: parameters.path,
-            type: 'GeoParquetSource',
-            model: this._model,
-          });
-
-          const geojsonData = Array.isArray(geojson) ? geojson[0] : geojson;
-
-          const format = new GeoJSON();
-
-          newSource = new VectorSource({
-            features: format.readFeatures(geojsonData, {
-              dataProjection: parameters.projection,
-              featureProjection: this._Map.getView().getProjection(),
-            }),
-          });
-          break;
-        }
-
-        case 'MarkerSource': {
-          const parameters = source.parameters as IMarkerSource;
-
-          const point = new Point(parameters.feature.coords);
-          const marker = new Feature({
-            type: 'icon',
-            geometry: point,
-          });
-
-          // Replace color placeholder in SVG with the parameter color
-          const markerColor = parameters.color || '#3463a0';
-          const svgString = markerIcon.svgstr
-            .replace('{{COLOR}}', markerColor)
-            .replace('<svg', '<svg width="128" height="128"');
-
-          const iconStyle = new Style({
-            image: new Icon({
-              src: `data:image/svg+xml;charset=utf-8,${encodeURIComponent(svgString)}`,
-              scale: 0.25,
-              anchor: [0.5, 1],
-              anchorXUnits: 'fraction',
-              anchorYUnits: 'fraction',
-            }),
-          });
-
-          marker.setStyle(iconStyle);
-
-          newSource = new VectorSource({
-            features: [marker],
-          });
-
-          break;
-        }
-
-        case 'WmsTileSource': {
-          const sourceParameters = source.parameters as IWmsTileSource;
-          const url = sourceParameters.url;
-          const selectedLayer = sourceParameters?.params?.layers;
-
-          newSource = new TileWMSSource({
-            attributions: sourceParameters?.attribution,
-            url,
-            params: {
-              LAYERS: selectedLayer,
-              TILED: true,
-            },
-          });
-
-          break;
-        }
-      }
-    } catch (err: any) {
-      this._log(
-        'error',
-        `Failed to load source "${source.name ?? id}" (${source.type}): ${err.message}`,
-      );
-      return;
-    }
-
-    this._log(
-      'info',
-      `Source "${source.name ?? id}" (${source.type}) loaded successfully`,
-    );
-    newSource.set('id', id);
-
-    // Forward OL tile/feature load errors to the JupyterLab log console.
-    // These errors (CORS failures, network errors, etc.) are written directly
-    // by the browser to DevTools and cannot be captured by console patching —
-    // OL's own events are the only reliable interception point.
-    newSource.on('tileloaderror', (evt: any) => {
-      const url = evt?.tile?.getKey?.() ?? '';
-      this._log(
-        'error',
-        `Tile load error for source "${id}"${url ? ': ' + url : ''}`,
-      );
-    });
-    newSource.on('featuresloaderror', () => {
-      this._log('error', `Features load error for source "${id}"`);
-    });
-
-    // _sources is a list of OpenLayers sources
-    this._sources[id] = newSource;
-
-    this._trackSourceExtZoom(id, newSource);
-  }
-
-  private computeSourceUrl(source: IJGISSource): string {
-    const parameters = source.parameters as IRasterSource;
-    const urlParameters = parameters.urlParameters || {};
-    let url: string = parameters.url;
-
-    for (const parameterName of Object.keys(urlParameters)) {
-      url = url.replace(`{${parameterName}}`, urlParameters[parameterName]);
-    }
-
-    // Special case for max_zoom and min_zoom
-    if (url.includes('{max_zoom}')) {
-      url = url.replace('{max_zoom}', parameters.maxZoom.toString());
-    }
-    if (url.includes('{min_zoom}')) {
-      url = url.replace('{min_zoom}', parameters.minZoom.toString());
-    }
-
-    return url;
-  }
-
-  /**
-   * Update a source in the map.
-   *
-   * @param id - the source id.
-   * @param source - the source object.
-   */
-  async updateSource(id: string, source: IJGISSource): Promise<void> {
-    // get the layer id associated with this source
-    const layerId = this._sourceToLayerMap.get(id);
-    // get the OL layer
-    const mapLayer = this.getLayer(layerId);
-    if (!mapLayer) {
-      return;
-    }
-    // remove source being updated
-    this.removeSource(id);
-    // create updated source
-    await this.addSource(id, source);
-    // change source of target layer
-    mapLayer.setSource(this._sources[id]);
-  }
-
-  /**
-   * Remove a source from the map.
-   *
-   * @param id - the source id.
-   */
-  removeSource(id: string): void {
-    delete this._sources[id];
-  }
-
-  /**
-   * Add or move the layers of the map.
-   *
-   * @param layerIds - the list of layers in the depth order (beneath first).
-   */
-  updateLayers(layerIds: string[]): void {
-    this._updateLayersImpl(layerIds);
-  }
-
-  /**
-   * Updates the position and existence of layers in the OL map based on the layer IDs.
-   *
-   * @param layerIds - An array of layer IDs that should be present on the map.
-   * @returns {} Nothing is returned.
-   */
-  private async _updateLayersImpl(layerIds: string[]): Promise<void> {
-    // get layers that are currently on the OL map
-    const previousLayerIds = this.getLayerIDs();
-
-    // Iterate over the new layer IDs:
-    //   * Add layers to the map that are present in the list but not the map.
-    //   * Remove layers from the map that are present in the map but not the list.
-    //   * Update layer positions to match the list.
-    for (
-      let targetLayerPosition = 0;
-      targetLayerPosition < layerIds.length;
-      targetLayerPosition++
-    ) {
-      const layerId = layerIds[targetLayerPosition];
-      const layer = this._model.sharedModel.getLayer(layerId);
-
-      if (this._loadingLayers.has(layerId)) {
-        continue;
-      }
-
-      if (!layer) {
-        this._log(
-          'warning',
-          `Layer with ID ${layerId} does not exist in the shared model.`,
-        );
-        continue;
-      }
-
-      const mapLayer = this.getLayer(layerId);
-
-      if (mapLayer !== undefined) {
-        this.moveLayer(layerId, targetLayerPosition);
-      } else {
-        await this.addLayer(layerId, layer, targetLayerPosition);
-      }
-
-      const previousIndex = previousLayerIds.indexOf(layerId);
-      if (previousIndex > -1) {
-        previousLayerIds.splice(previousIndex, 1);
-      }
-    }
-
-    // Remove layers that are no longer in the `layerIds` list.
-    previousLayerIds.forEach(layerId => {
-      const layer = this.getLayer(layerId);
-      if (layer !== undefined) {
-        this._Map.removeLayer(layer);
-      }
-    });
-
-    this._ready = true;
-
-    // If a "zoom to layer" request arrived before its layer was on the map,
-    // retry now that layers have been (re)built.
-    if (
-      this._pendingZoomLayerId &&
-      this.getLayer(this._pendingZoomLayerId) !== undefined
-    ) {
-      const pendingId = this._pendingZoomLayerId;
-      this._pendingZoomLayerId = null;
-      this._onZoomToPosition(this._model, pendingId);
-    }
-  }
-
-  /**
-   * Build the map layer.
-   *
-   * @param id - id of the layer.
-   * @param layer - the layer object.
-   * @returns - the map layer.
-   */
-  private async _buildMapLayer(
-    id: string,
-    layer: IJGISLayer,
-  ): Promise<Layer | LayerGroup | StacLayer | undefined> {
-    this.setState(old => ({ ...old, loadingLayer: true }));
-    this._loadingLayers.add(id);
-
-    let newMapLayer: OlLayerTypes;
-    let layerParameters: any;
-    let sourceId: string | undefined;
-    let source: IJGISSource | undefined;
-
-    // Sourceless layers
-    if (!['StacLayer', 'StorySegmentLayer'].includes(layer.type)) {
-      sourceId = layer.parameters?.source;
-      if (!sourceId) {
-        return;
-      }
-      source = this._model.sharedModel.getLayerSource(sourceId);
-      if (!source) {
-        return;
-      }
-      if (!this._sources[sourceId]) {
-        await this.addSource(sourceId, source);
-      }
-      // Updates GeoZarr layer again after getting bands, styles info from zarr store.
-      source = this._model.sharedModel.getLayerSource(sourceId);
-    }
-
-    // TODO: OpenLayers provides a bunch of sources for specific tile
-    // providers, so maybe set up some way to use those
-    switch (layer.type) {
-      case 'RasterLayer': {
-        layerParameters = layer.parameters as IRasterLayer;
-
-        newMapLayer = new TileLayer({
-          opacity: layerParameters.opacity,
-          visible: layer.visible,
-          source: this._sources[layerParameters.source],
-        });
-
-        break;
-      }
-      case 'VectorLayer': {
-        layerParameters = layer.parameters as IVectorLayer;
-
-        if (Array.isArray(layerParameters.symbologyState?.layers)) {
-          const olSource = this._sources[
-            layerParameters.source
-          ] as VectorSource;
-          const grammarState =
-            layerParameters.symbologyState as IGrammarSymbologyState;
-          const rows =
-            olSource instanceof VectorSource
-              ? olSource.getFeatures().map(f => (f as Feature).getProperties())
-              : [];
-          const featureValues = extractEncodingFieldValues(grammarState, rows);
-          newMapLayer = grammarToOLLayer(
-            layerParameters.symbologyState as IGrammarSymbologyState,
-            olSource,
-            layerParameters.opacity,
-            layer.visible,
-            featureValues,
-          ) as OlLayerTypes;
-        } else {
-          newMapLayer = new VectorImageLayer({
-            opacity: layerParameters.opacity,
-            visible: layer.visible,
-            source: this._sources[layerParameters.source],
-            style: this.vectorLayerStyleRuleBuilder(layer),
-          });
-        }
-
-        break;
-      }
-      case 'VectorTileLayer': {
-        layerParameters = layer.parameters as IVectorLayer;
-
-        newMapLayer = new VectorTileLayer({
-          opacity: layerParameters.opacity,
-          visible: layer.visible,
-          source: this._sources[layerParameters.source],
-          style: this.vectorLayerStyleRuleBuilder(layer),
-        });
-
-        break;
-      }
-      case 'HillshadeLayer': {
-        layerParameters = layer.parameters as IHillshadeLayer;
-
-        newMapLayer = new RasterLayer({
-          opacity: layerParameters.opacity ?? 0.3,
-          visible: layer.visible,
-          source: this._sources[layerParameters.source],
-          style: {
-            color: ['color', this.hillshadeMath()],
-          },
-        });
-
-        break;
-      }
-      case 'ImageLayer': {
-        layerParameters = layer.parameters as IImageLayer;
-
-        newMapLayer = new ImageLayer({
-          opacity: layerParameters.opacity,
-          visible: layer.visible,
-          source: this._sources[layerParameters.source],
-        });
-
-        break;
-      }
-      case 'OpenEOTileLayer': {
-        layerParameters = layer.parameters as IOpenEOTileLayer;
-
-        newMapLayer = new OpenEOTileLayer({
-          opacity: layerParameters.opacity,
-          visible: layer.visible,
-          source: this._sources[layerParameters.source],
-        });
-        break;
-      }
-      case 'GeoTiffLayer': {
-        layerParameters = layer.parameters as IGeoTiffLayer;
-        const geoTiffSource = this._sources[layerParameters.source];
-
-        await this._waitForSourceReady(geoTiffSource);
-
-        if (Array.isArray(layerParameters.symbologyState?.layers)) {
-          newMapLayer = grammarToOLLayer(
-            layerParameters.symbologyState as IGrammarSymbologyState,
-            geoTiffSource,
-            layerParameters.opacity ?? 1,
-            layer.visible ?? true,
-            [],
-            true,
-          ) as OlLayerTypes;
-        } else {
-          // This is to handle python sending a None for the color
-          const layerOptions: any = {
-            opacity: layerParameters.opacity,
-            visible: layer.visible,
-            source: geoTiffSource,
-          };
-
-          if (layerParameters.color) {
-            layerOptions['style'] = {
-              color: layerParameters.color,
-            };
-          }
-
-          newMapLayer = new RasterLayer(layerOptions);
-        }
-        break;
-      }
-      case 'GeoZarrLayer': {
-        layerParameters = layer.parameters as IGeoZarrLayer;
-        const geoZarrSource = this._sources[layerParameters.source];
-
-        await this._waitForSourceReady(geoZarrSource);
-
-        if (Array.isArray(layerParameters.symbologyState?.layers)) {
-          newMapLayer = grammarToOLLayer(
-            layerParameters.symbologyState as IGrammarSymbologyState,
-            geoZarrSource,
-            layerParameters.opacity ?? 1,
-            layer.visible ?? true,
-            [],
-            true,
-          ) as OlLayerTypes;
-        } else {
-          const bands = source?.parameters?.bands || [];
-          const defaultColor = buildZarrColorStyle(bands);
-
-          newMapLayer = new RasterLayer({
-            opacity: layerParameters.opacity ?? 1,
-            visible: layer.visible,
-            source: geoZarrSource,
-            style: {
-              gamma: layerParameters.gamma ?? 1,
-              color: layerParameters.color ?? defaultColor,
-            },
-          });
-        }
-
-        break;
-      }
-
-      case 'StacLayer': {
-        layerParameters = layer.parameters as IStacLayer;
-
-        newMapLayer = new StacLayer({
-          displayPreview: true,
-          data: layerParameters.data,
-          opacity: layerParameters.opacity,
-          visible: layer.visible,
-          assets: Object.keys(layerParameters.data.assets),
-          extent: layerParameters.data.bbox,
-        });
-
-        this.setState(old => ({
-          ...old,
-          metadata: layerParameters.data.properties,
-        }));
-
-        break;
-      }
-
-      case 'StorySegmentLayer': {
-        // Special layer not for this
-        return;
-      }
-    }
-
-    // OpenLayers doesn't have name/id field so add it
-    newMapLayer.set('id', id);
-
-    // STAC layers don't have source
-    if (newMapLayer instanceof Layer) {
-      // we need to keep track of which source has which layers
-      // Only set sourceToLayerMap if 'source' exists on layerParameters
-      if ('source' in layerParameters) {
-        this._sourceToLayerMap.set(layerParameters.source, id);
-      }
-
-      this.addProjection(newMapLayer);
-      await this._waitForLayerReady(newMapLayer);
-    }
-
-    this._loadingLayers.delete(id);
-    return newMapLayer;
-  }
-
-  private ensureProjectionRegistered(projectionCode: string): void {
-    const hasProj4Definition = Boolean(proj4.defs(projectionCode));
-    if (!hasProj4Definition) {
-      // Check if the projection exists in proj4list
-      const proj4Definition = proj4list[projectionCode];
-      if (!proj4Definition) {
-        this._log(
-          'warning',
-          `Projection code '${projectionCode}' not found in proj4list`,
-        );
-        return;
-      }
-
-      try {
-        if (Array.isArray(proj4Definition)) {
-          proj4.defs([proj4Definition]);
-        } else {
-          proj4.defs(projectionCode, proj4Definition);
-        }
-      } catch (error: any) {
-        this._log(
-          'warning',
-          `Failed to register projection '${projectionCode}'. Error: ${error.message}`,
-        );
-        return;
-      }
-    }
-
-    // Always register after ensuring proj4 defs exist so OL transform functions
-    // are available even when the projection object already exists in cache.
-    register(proj4 as any);
-  }
-
-  private resolveLayerSourceProjection(layer: IJGISLayer): string | undefined {
-    const sourceId = layer.parameters?.source;
-    if (!sourceId) {
-      return;
-    }
-
-    const source = this._model.sharedModel.getLayerSource(sourceId);
-    if (!source) {
-      return;
-    }
-
-    const parameters = source.parameters;
-    return parameters?.projection;
-  }
-
-  addProjection(target: Layer | IJGISLayer): void {
-    if (target instanceof Layer) {
-      const sourceProjection = target.getSource()?.getProjection();
-      if (!sourceProjection) {
-        this._log('warning', 'Layer source projection is undefined or invalid');
-        return;
-      }
-
-      this.ensureProjectionRegistered(sourceProjection.getCode());
-      return;
-    }
-
-    const projectionCode = this.resolveLayerSourceProjection(target);
-    if (!projectionCode) {
-      return;
-    }
-
-    this.ensureProjectionRegistered(projectionCode);
-  }
-
-  /**
-   * Add a layer to the map.
-   *
-   * @param id - id of the layer.
-   * @param layer - the layer object.
-   * @param index - expected index of the layer.
-   */
-  async addLayer(id: string, layer: IJGISLayer, index: number): Promise<void> {
-    if (this.getLayer(id)) {
-      // Layer already exists
-      return;
-    }
-
-    try {
-      this.addProjection(layer);
-      const newMapLayer = await this._buildMapLayer(id, layer);
-      if (newMapLayer !== undefined) {
-        await this._waitForReady();
-
-        // Adjust index to ensure it's within bounds
-        const numLayers = this._Map.getLayers().getLength();
-        const safeIndex = Math.min(index, numLayers);
-        this._Map.getLayers().insertAt(safeIndex, newMapLayer);
-        this._trackLayerViewState(id, newMapLayer);
-
-        // doing +1 instead of calling method again
-        if (
-          !this.state.initialLayersReady &&
-          numLayers + 1 === this._initialLayersCount
-        ) {
-          this.setState(old => ({ ...old, initialLayersReady: true }));
-        }
-      }
-
-      if (layer.type !== 'StorySegmentLayer') {
-        this._model.syncSelected(
-          { [id]: { type: 'layer' } },
-          this._model.getClientId().toString(),
-        );
-      }
-    } catch (error: any) {
-      if (
-        this.state.loadingErrors.find(
-          item => item.id === id && item.error === error.message,
-        )
-      ) {
-        return;
-      }
-
-      await showErrorMessage(
-        `Error Adding ${layer.name}`,
-        `Failed to add ${layer.name}: ${error.message || 'invalid file path'}`,
-      );
-      this.setState(old => ({ ...old, loadingLayer: false }));
-      this.state.loadingErrors.push({
-        id,
-        error: error.message || 'invalid file path',
-        index,
-      });
-    } finally {
-      this._loadingLayers.delete(id);
-      this.setState(old => ({ ...old, loadingLayer: false }));
-    }
-  }
-
-  // Used by VectorTileLayer (which shares a flat-style API with Grammar output).
-  vectorLayerStyleRuleBuilder = (layer: IJGISLayer) => {
-    const layerParams = layer.parameters as IVectorLayer | undefined;
-    const ss = layerParams?.symbologyState;
-    if (!ss || Object.keys(ss).length === 0) {
-      return [{ style: DEFAULT_FLAT_STYLE }];
-    }
-
-    const flatStyle = grammarToOLStyle(
-      layerParams.symbologyState as IGrammarSymbologyState,
-      [],
-    );
-
-    const layerStyle: Rule = { style: flatStyle };
-
-    if (layer.filters?.logicalOp && layer.filters.appliedFilters?.length > 0) {
-      const buildCondition = (filter: IJGISFilterItem): any[] => {
-        const base = [filter.operator, ['get', filter.feature]];
-        return filter.operator === 'between'
-          ? [...base, filter.betweenMin, filter.betweenMax]
-          : [...base, filter.value];
-      };
-
-      layerStyle.filter =
-        layer.filters.appliedFilters.length === 1
-          ? buildCondition(layer.filters.appliedFilters[0])
-          : [
-              layer.filters.logicalOp,
-              ...layer.filters.appliedFilters.map(buildCondition),
-            ];
-    }
-
-    return [layerStyle];
-  };
-
-  /**
-   * Taken from https://openlayers.org/en/latest/examples/webgl-shaded-relief.html
-   * @returns
-   */
-  private hillshadeMath = () => {
-    // The method used to extract elevations from the DEM.
-    // In this case the format used is Terrarium
-    // red * 256 + green + blue / 256 - 32768
-    //
-    // Other frequently used methods include the Mapbox format
-    // (red * 256 * 256 + green * 256 + blue) * 0.1 - 10000
-    //
-    function elevation(xOffset: number, yOffset: number) {
-      const red = ['band', 1, xOffset, yOffset];
-      const green = ['band', 2, xOffset, yOffset];
-      const blue = ['band', 3, xOffset, yOffset];
-
-      // band math operates on normalized values from 0-1
-      // so we scale by 255
-      return [
-        '+',
-        ['*', 255 * 256, red],
-        ['*', 255, green],
-        ['*', 255 / 256, blue],
-        -32768,
-      ];
-    }
-    // Generates a shaded relief image given elevation data.  Uses a 3x3
-    // neighborhood for determining slope and aspect.
-    const dp = ['*', 2, ['resolution']];
-    const z0x = ['*', 2, elevation(-1, 0)];
-    const z1x = ['*', 2, elevation(1, 0)];
-    const dzdx = ['/', ['-', z1x, z0x], dp];
-    const z0y = ['*', 2, elevation(0, -1)];
-    const z1y = ['*', 2, elevation(0, 1)];
-    const dzdy = ['/', ['-', z1y, z0y], dp];
-    const slope = ['atan', ['sqrt', ['+', ['^', dzdx, 2], ['^', dzdy, 2]]]];
-    const aspect = ['clamp', ['atan', ['-', 0, dzdx], dzdy], -Math.PI, Math.PI];
-    const sunEl = ['*', Math.PI / 180, 45];
-    const sunAz = ['*', Math.PI / 180, 46];
-
-    const cosIncidence = [
-      '+',
-      ['*', ['sin', sunEl], ['cos', slope]],
-      ['*', ['cos', sunEl], ['sin', slope], ['cos', ['-', sunAz, aspect]]],
-    ];
-    const scaled = ['*', 255, cosIncidence];
-
-    return scaled;
-  };
-
-  private _syncGrammarSubLayers(
-    id: string,
-    layer: IJGISLayer,
-    mapLayer: Layer | LayerGroup,
-  ): void {
-    const layerParams = layer.parameters as
-      | IVectorLayer
-      | IGeoTiffLayer
-      | IGeoZarrLayer
-      | undefined;
-    const grammarState = layerParams?.symbologyState as
-      | IGrammarSymbologyState
-      | undefined;
-
-    if (!grammarState || !Array.isArray(grammarState.layers)) {
-      return;
-    }
-
-    const sourceId = layerParams?.source;
-    const source = sourceId ? this._sources[sourceId] : undefined;
-    const rows =
-      source instanceof VectorSource
-        ? source.getFeatures().map(f => (f as Feature).getProperties())
-        : [];
-    const featureValues = extractEncodingFieldValues(grammarState, rows);
-    const nextLayer = grammarToOLLayer(
-      grammarState,
-      source,
-      layerParams?.opacity ?? layer.parameters?.opacity ?? 1,
-      layer.visible,
-      featureValues,
-      layer.type === 'GeoTiffLayer' || layer.type === 'GeoZarrLayer',
-    );
-
-    if (mapLayer instanceof LayerGroup) {
-      if (nextLayer instanceof LayerGroup) {
-        const replacementLayers = nextLayer.getLayers().getArray();
-        mapLayer.setOpacity(
-          layerParams?.opacity ?? layer.parameters?.opacity ?? 1,
-        );
-        mapLayer.setVisible(layer.visible);
-        mapLayer.setLayers(new Collection(replacementLayers));
-        return;
-      }
-
-      // Collapse back to a single top-level layer so tools that expect a
-      // concrete OL Layer (fly-to/identify) keep working.
-      nextLayer.set('id', id);
-      const index = this.getLayerIndex(id);
-      if (index !== -1) {
-        this._Map.getLayers().setAt(index, nextLayer);
-      }
-      return;
-    }
-
-    nextLayer.set('id', id);
-    const index = this.getLayerIndex(id);
-    if (index !== -1) {
-      this._Map.getLayers().setAt(index, nextLayer);
-    }
-  }
-
-  /**
-   * Update a layer of the map.
-   *
-   * Only for updating appearance -- opacity or style.
-   * For vector layers, the whole layer is replaced.
-   *
-   * @param id - id of the layer.
-   * @param layer - the layer object.
-   */
-  async updateLayer(
-    id: string,
-    layer: IJGISLayer,
-    mapLayer: Layer,
-    oldLayer?: IDict,
-  ): Promise<void> {
-    layer.type !== 'StorySegmentLayer' && mapLayer.setVisible(layer.visible);
-
-    switch (layer.type) {
-      case 'RasterLayer': {
-        mapLayer.setOpacity(layer.parameters?.opacity ?? 1);
-        break;
-      }
-      case 'VectorLayer': {
-        const layerParams = layer.parameters as IVectorLayer;
-
-        if (Array.isArray(layerParams.symbologyState?.layers)) {
-          this._syncGrammarSubLayers(id, layer, mapLayer as Layer | LayerGroup);
-          break;
-        }
-
-        mapLayer.setOpacity(layerParams.opacity ?? 1);
-
-        (mapLayer as VectorImageLayer).setStyle(
-          this.vectorLayerStyleRuleBuilder(layer),
-        );
-
-        break;
-      }
-      case 'VectorTileLayer': {
-        const layerParams = layer.parameters as IVectorTileLayer;
-
-        mapLayer.setOpacity(layerParams.opacity ?? 1);
-
-        (mapLayer as VectorTileLayer).setStyle(
-          this.vectorLayerStyleRuleBuilder(layer),
-        );
-
-        break;
-      }
-      case 'HillshadeLayer': {
-        // TODO figure out color here
-        mapLayer.setOpacity(layer.parameters?.opacity ?? 0.3);
-        break;
-      }
-      case 'ImageLayer': {
-        mapLayer.setOpacity(layer.parameters?.opacity ?? 1);
-        break;
-      }
-      case 'GeoTiffLayer': {
-        if (Array.isArray(layer?.parameters?.symbologyState?.layers)) {
-          this._syncGrammarSubLayers(id, layer, mapLayer as Layer | LayerGroup);
-        } else {
-          mapLayer.setOpacity(layer.parameters?.opacity);
-          if (layer?.parameters?.color) {
-            (mapLayer as RasterLayer).setStyle({
-              color: layer.parameters.color,
-            });
-          }
-        }
-        break;
-      }
-      case 'GeoZarrLayer': {
-        if (Array.isArray(layer?.parameters?.symbologyState?.layers)) {
-          this._syncGrammarSubLayers(id, layer, mapLayer as Layer | LayerGroup);
-          break;
-        }
-
-        if (layer?.parameters?.opacity !== undefined) {
-          mapLayer.setOpacity(layer.parameters?.opacity);
-        }
-
-        const source = this._model.getSource(layer?.parameters?.source);
-        const bands = (source?.parameters as any)?.bands || [];
-
-        const style: any = {
-          color: layer.parameters?.color ?? buildZarrColorStyle(bands),
-          gamma: layer.parameters?.gamma ?? 1,
-        };
-
-        (mapLayer as RasterLayer).setStyle(style);
-
-        break;
-      }
-      case 'OpenEOTileLayer': {
-        const layerParams = layer.parameters as IOpenEOTileLayer;
-        const openeoLayer = mapLayer as OpenEOTileLayer;
-
-        openeoLayer.setOpacity(layerParams.opacity ?? 1);
-
-        break;
-      }
-      case 'StacLayer':
-        mapLayer.setOpacity(layer.parameters?.opacity ?? 1);
-        break;
-    }
-  }
-
   private flyToGeometry(sender: IJupyterGISModel, geometry: any): void {
     if (!geometry || typeof geometry.getExtent !== 'function') {
       this._log('warning', `Invalid geometry for flyToGeometry: ${geometry}`);
@@ -2390,197 +1037,6 @@ export class MainView extends React.Component<IMainViewProps, IStates> {
 
   private _buildHighlightStyle(original: Style, geomType?: string): Style {
     return buildHighlightStyle(original, geomType);
-  }
-
-  /**
-   * Compute extent for layer or source
-   */
-  private _computeExtent(
-    layer?: Layer | StacLayer,
-    source?: any,
-  ): number[] | undefined {
-    try {
-      if (source instanceof VectorSource) {
-        const extent = source.getExtent();
-        if (extent) {
-          return extent;
-        }
-      }
-
-      if (source instanceof TileSource || source instanceof VectorTileSource) {
-        const tileGrid = source.getTileGrid();
-        const extent = tileGrid?.getExtent();
-        if (extent) {
-          return extent;
-        }
-      }
-
-      if (layer instanceof StacLayer) {
-        const extent = layer.getExtent();
-        if (extent) {
-          return extent;
-        }
-      }
-    } catch (error) {
-      this._log('warning', `Failed to compute extent: ${error}`);
-    }
-
-    return undefined;
-  }
-
-  private _computeZoomFromExtent(extent: number[]): number | null {
-    if (!this._Map) {
-      return null;
-    }
-
-    const view = this._Map.getView();
-    const size = this._Map.getSize() ?? getSize(extent);
-
-    const resolution = view.getResolutionForExtent(extent, size);
-    const zoom = view.getZoomForResolution(resolution);
-
-    return zoom ?? view.getZoom() ?? 0;
-  }
-
-  /**
-   * Track layer's extent and zoom in model's view state
-   */
-  private _trackLayerViewState(
-    layerId: string,
-    olLayer: Layer | LayerGroup,
-  ): void {
-    const effectiveLayer =
-      olLayer instanceof LayerGroup
-        ? (olLayer.getLayers().getArray()[0] as Layer | undefined)
-        : olLayer;
-    if (!effectiveLayer) {
-      return;
-    }
-    const source = effectiveLayer.getSource();
-    const sourceId = source?.get?.('id');
-
-    let extent = sourceId ? this._model.getExtent(sourceId) : undefined;
-
-    if (!extent) {
-      extent = this._computeExtent(effectiveLayer, source);
-    }
-
-    if (extent) {
-      const zoom = this._computeZoomFromExtent(extent);
-
-      if (zoom === null) {
-        return;
-      }
-
-      const view: IViewState[string] = { extent, zoom };
-      this._model.updateLayerViewState(layerId, view);
-    }
-  }
-
-  /**
-   * Track source's extent and zoom in model's view state
-   */
-  private _trackSourceExtZoom(sourceId: string, olSource: Source): void {
-    const extent = this._computeExtent(undefined, olSource);
-
-    if (extent) {
-      const projection = olSource?.getProjection?.()?.getCode?.();
-      const zoom = this._computeZoomFromExtent(extent);
-
-      if (zoom === null) {
-        return;
-      }
-
-      const view: IViewState[string] = {
-        extent,
-        zoom,
-        ...(projection && { projection }),
-      };
-      this._model.updateLayerViewState(sourceId, view);
-    }
-  }
-
-  /**
-   * Wait for all layers to be loaded.
-   */
-  private _waitForReady(): Promise<void> {
-    return new Promise(resolve => {
-      const checkReady = () => {
-        if (this._loadingLayers.size === 0) {
-          this.setState(old => ({
-            ...old,
-            loadingLayer: false,
-          }));
-          resolve();
-        } else {
-          setTimeout(checkReady, 50);
-        }
-      };
-
-      checkReady();
-    });
-  }
-
-  /**
-   * Wait for a layers source state to be 'ready'
-   * @param layer The Layer to check
-   */
-  private _waitForLayerReady(layer: Layer | LayerGroup) {
-    return new Promise<void>((resolve, reject) => {
-      const checkState = () => {
-        const state = layer.getSourceState();
-        if (state === 'ready') {
-          layer.un('change', checkState);
-          resolve();
-        } else if (state === 'error') {
-          layer.un('change', checkState);
-          reject(new Error('Layer failed to load.'));
-        }
-      };
-
-      // Listen for state changes
-      layer.on('change', checkState);
-
-      // Check the state immediately in case it's already 'ready'
-      checkState();
-    });
-  }
-
-  /**
-   * Wait for a source to finish loading and configuring
-   * @param source The Source to check
-   */
-  private _waitForSourceReady(source: Source) {
-    return new Promise<void>((resolve, reject) => {
-      const checkState = () => {
-        const state = source.getState();
-        if (state === 'ready') {
-          source.un('change', checkState);
-          resolve();
-        } else if (state === 'error') {
-          source.un('change', checkState);
-          reject(new Error('Source failed to load'));
-        }
-      };
-
-      // Listen for state changes
-      source.on('change', checkState);
-
-      // Check immediately in case already ready
-      checkState();
-    });
-  }
-
-  /**
-   * Remove a layer from the map.
-   *
-   * @param id - the id of the layer.
-   */
-  removeLayer(id: string): void {
-    const mapLayer = this.getLayer(id);
-    if (mapLayer) {
-      this._Map.removeLayer(mapLayer);
-    }
   }
 
   private _handleSelectedChanged = (): void => {
@@ -2868,7 +1324,6 @@ export class MainView extends React.Component<IMainViewProps, IStates> {
       // Dispose old viewer
       if (this._mapViewer) {
         this._mapViewer.destroy();
-        this._mapViewer = null;
       }
 
       // Get current view state before regenerating
@@ -2879,9 +1334,9 @@ export class MainView extends React.Component<IMainViewProps, IStates> {
       // Regenerate map with new viewer
       await this.generateMap(center, zoom, projection);
 
-      const layerids = this.getLayerIDs();
+      const layerids = this._mapViewer.getLayerIDs();
 
-      this.updateLayers(layerids);
+      this._mapViewer.updateLayers(layerids);
     } catch (error) {
       console.error('Error regenerating map:', error);
     }
@@ -2948,17 +1403,6 @@ export class MainView extends React.Component<IMainViewProps, IStates> {
   }
 
   /**
-   * Convenience method to get a specific layer from OpenLayers Map
-   * @param id Layer to retrieve
-   */
-  private getLayer(id: string) {
-    return this._Map
-      .getLayers()
-      .getArray()
-      .find(layer => layer.get('id') === id) as Layer;
-  }
-
-  /**
    * Convenience method to get a specific layer index from OpenLayers Map
    * @param id Layer to retrieve
    */
@@ -2967,42 +1411,6 @@ export class MainView extends React.Component<IMainViewProps, IStates> {
       .getLayers()
       .getArray()
       .findIndex(layer => layer.get('id') === id);
-  }
-
-  /**
-   * Convenience method to get list layer IDs from the OpenLayers Map
-   */
-  private getLayerIDs(): string[] {
-    return this._Map
-      .getLayers()
-      .getArray()
-      .map(layer => layer.get('id'));
-  }
-
-  /**
-   * Move layer `id` in the stack to `index`.
-   *
-   * @param id - id of the layer.
-   * @param index - expected index of the layer.
-   */
-  moveLayer(id: string, index: number): void {
-    const currentIndex = this.getLayerIndex(id);
-    if (currentIndex === index || currentIndex === -1) {
-      return;
-    }
-    const layer = this.getLayer(id);
-
-    // should not be undefined since the id exists above
-    if (layer === undefined) {
-      return;
-    }
-    this._Map.getLayers().removeAt(currentIndex);
-
-    // Adjust index to ensure it's within bounds
-    const numLayers = this._Map.getLayers().getLength();
-    const safeIndex = Math.min(index, numLayers);
-
-    this._Map.getLayers().insertAt(safeIndex, layer);
   }
 
   private _onLayersChanged(
@@ -3019,7 +1427,7 @@ export class MainView extends React.Component<IMainViewProps, IStates> {
       const { id, oldValue: oldLayer, newValue: newLayer } = change;
 
       if (!newLayer || Object.keys(newLayer).length === 0) {
-        this.removeLayer(id);
+        this._mapViewer.removeLayer(id);
         if (
           this._model.currentMode === 'drawing' &&
           this._model.checkIfIsADrawVectorLayer(oldLayer as IJGISLayer)
@@ -3030,17 +1438,17 @@ export class MainView extends React.Component<IMainViewProps, IStates> {
         return;
       }
 
-      const mapLayer = this.getLayer(id);
+      const mapLayer = this._mapViewer.getLayer(id);
       const layerTree = JupyterGISModel.getOrderedLayerIds(this._model);
 
       if (layerTree.includes(id)) {
-        this.updateLayer(id, newLayer, mapLayer, oldLayer);
+        this._mapViewer.updateLayer(id, newLayer, mapLayer, oldLayer);
 
         if (mapLayer) {
-          this._trackLayerViewState(id, mapLayer);
+          this._mapViewer.trackLayerViewState(id, mapLayer);
         }
       } else {
-        this.updateLayers(layerTree);
+        this._mapViewer.updateLayers(layerTree);
       }
     });
   }
@@ -3052,7 +1460,9 @@ export class MainView extends React.Component<IMainViewProps, IStates> {
     this._ready = false;
     // We can't properly use the change, because of the nested groups in the the shared
     // document which is flattened for the map tool.
-    this.updateLayers(JupyterGISModel.getOrderedLayerIds(this._model));
+    this._mapViewer.updateLayers(
+      JupyterGISModel.getOrderedLayerIds(this._model),
+    );
   }
 
   /**
@@ -3088,7 +1498,7 @@ export class MainView extends React.Component<IMainViewProps, IStates> {
       if (normalize(sourceServerUrl) === normalize(serverUrl)) {
         // updateSource removes the OL source and reconstructs it; the new
         // OpenEOTileSource finds the cached connection and renders.
-        void this.updateSource(sourceId, source);
+        void this._mapViewer.updateSource(sourceId, source);
       }
     }
   }
@@ -3103,11 +1513,11 @@ export class MainView extends React.Component<IMainViewProps, IStates> {
 
     change.sourceChange?.forEach(change => {
       if (!change.newValue || Object.keys(change.newValue).length === 0) {
-        this.removeSource(change.id);
+        this._mapViewer.removeSource(change.id);
       } else {
         const source = this._model.getSource(change.id);
         if (source) {
-          this.updateSource(change.id, source);
+          this._mapViewer.updateSource(change.id, source);
         }
       }
     });
@@ -3477,124 +1887,6 @@ export class MainView extends React.Component<IMainViewProps, IStates> {
     );
   }
 
-  // TODO this and flyToPosition need a rework
-  private _onZoomToPosition(_: IJupyterGISModel, id: string) {
-    // Check if the id is an annotation
-    const annotation = this._model.annotationModel?.getAnnotation(id);
-    if (annotation) {
-      this._flyToPosition(annotation.position, annotation.zoom);
-      return;
-    }
-
-    // The id is a layer
-    const layer = this.getLayer(id);
-    const source = layer?.getSource();
-    const jgisLayer = this._model.getLayer(id);
-
-    /**
-     * Layer may be undefined in two cases:
-     * 1. StorySegmentLayer: These layers don't have an associated OpenLayers layer
-     * 2. StacLayer: When centerOnPosition is called immediately after adding the layer,
-     *    the OpenLayers layer hasn't been created yet, so we use the bbox from the
-     *    layer model's STAC data directly.
-     */
-    if (!layer) {
-      // Handle StacLayer that hasn't been added to the map yet
-      if (jgisLayer?.type === 'StacLayer') {
-        const layerParams = jgisLayer.parameters as IStacLayer;
-        const stacBbox = layerParams.data?.bbox;
-
-        if (stacBbox && stacBbox.length === 4) {
-          // STAC bbox format: [west, south, east, north] in EPSG:4326
-          const [west, south, east, north] = stacBbox;
-          const bboxExtent = [west, south, east, north];
-
-          // Convert from EPSG:4326 to view projection
-          const viewProjection = this._Map.getView().getProjection();
-          const transformedExtent =
-            viewProjection.getCode() !== 'EPSG:4326'
-              ? transformExtent(bboxExtent, 'EPSG:4326', viewProjection)
-              : bboxExtent;
-
-          this._Map.getView().fit(transformedExtent, {
-            size: this._Map.getSize(),
-            duration: 500,
-            padding: [250, 250, 250, 250],
-          });
-          return;
-        }
-      }
-
-      // Handle StorySegmentLayer
-      if (jgisLayer?.type === 'StorySegmentLayer') {
-        const layerParams = jgisLayer.parameters as IStorySegmentLayer;
-        const coords = getCenter(layerParams.extent);
-
-        // Don't move map if we're already centered on the segment
-        const viewCenter = this._Map.getView().getCenter();
-        const centersEqual =
-          viewCenter !== undefined &&
-          Math.abs(viewCenter[0] - coords[0]) < 1e-9 &&
-          Math.abs(viewCenter[1] - coords[1]) < 1e-9;
-        if (centersEqual) {
-          return;
-        }
-
-        this._flyToPosition(
-          { x: coords[0], y: coords[1] },
-          layerParams.zoom,
-          (layerParams.transition.time ?? 1) * 1000, // seconds -> ms
-          layerParams.transition.type,
-        );
-
-        return;
-      }
-
-      // Generic layer whose OpenLayers layer hasn't been created yet (e.g. a
-      // layer just added via the Python API with zoom_to=True). Remember the
-      // request and retry once the layer has been added to the map.
-      if (jgisLayer) {
-        this._pendingZoomLayerId = id;
-        return;
-      }
-    }
-
-    const extent = this._computeExtent(layer, source);
-    if (!extent) {
-      this._log('warning', 'Layer ${id} has no extent.');
-      return;
-    }
-
-    if (!extent.every(value => Number.isFinite(value))) {
-      this._log(
-        'warning',
-        `Layer ${id} has an invalid extent: ${extent.join(', ')}`,
-      );
-      return;
-    }
-
-    // Convert layer extent value to view projection if needed
-    const sourceProjection = source?.getProjection();
-    const viewProjection = this._Map.getView().getProjection();
-
-    const transformedExtent =
-      sourceProjection && sourceProjection !== viewProjection
-        ? transformExtent(extent, sourceProjection, viewProjection)
-        : extent;
-    if (!transformedExtent.every(value => Number.isFinite(value))) {
-      this._log(
-        'warning',
-        `Layer ${id} has an invalid transformed extent: ${transformedExtent.join(', ')}`,
-      );
-      return;
-    }
-
-    this._Map.getView().fit(transformedExtent, {
-      size: this._Map.getSize(),
-      duration: 500,
-    });
-  }
-
   private _moveToPosition(
     center: { x: number; y: number },
     zoom: number,
@@ -3612,61 +1904,6 @@ export class MainView extends React.Component<IMainViewProps, IStates> {
         duration,
       });
     }
-  }
-
-  private _flyToPosition(
-    center: { x: number; y: number },
-    zoom: number,
-    duration = 1000,
-    transitionType?: 'linear' | 'immediate' | 'smooth',
-  ) {
-    const view = this._Map.getView();
-
-    // Cancel any in-progress animations before starting new ones
-    view.cancelAnimations();
-
-    const targetCenter: Coordinate = [center.x, center.y];
-
-    if (transitionType === 'linear') {
-      // Linear: direct zoom
-      view.animate({
-        center: targetCenter,
-        zoom: zoom,
-        duration,
-      });
-
-      return;
-    }
-
-    if (transitionType === 'smooth') {
-      // Smooth: zoom out, center, and zoom in
-      // Centering takes full duration, zoom out completes halfway, zoom in starts halfway
-      // 3 shows most of the map
-      const zoomOutLevel = 3;
-
-      // Start centering (full duration) and zoom out (50% duration) simultaneously
-      view.animate({
-        center: targetCenter,
-        duration: duration,
-      });
-      // Chain zoom out -> zoom in (zoom in starts when zoom out completes)
-      view.animate(
-        {
-          zoom: zoomOutLevel,
-          duration: duration * 0.5,
-        },
-        {
-          zoom: zoom,
-          duration: duration * 0.5,
-        },
-      );
-
-      return;
-    }
-
-    // Immediate move
-    view.setCenter(targetCenter);
-    view.setZoom(zoom);
   }
 
   private _lastPointerCoord: Coordinate | null = null;
@@ -3718,10 +1955,14 @@ export class MainView extends React.Component<IMainViewProps, IStates> {
     };
 
     this._model.sharedModel.addSource(sourceId, sourceModel);
-    await this.addSource(sourceId, sourceModel);
+    await this._mapViewer.addSource(sourceId, sourceModel);
 
     this._model.addLayer(layerId, layerModel);
-    await this.addLayer(layerId, layerModel, this.getLayerIDs().length);
+    await this._mapViewer.addLayer(
+      layerId,
+      layerModel,
+      this._mapViewer.getLayerIDs().length,
+    );
   }
 
   private _identifyFeature(e: MapBrowserEvent<any>) {
@@ -3816,7 +2057,7 @@ export class MainView extends React.Component<IMainViewProps, IStates> {
 
       case 'GeoTiffLayer':
       case 'GeoZarrLayer': {
-        const layer = this.getLayer(layerId) as RasterLayer;
+        const layer = this._mapViewer.getLayer(layerId) as RasterLayer;
         const data = layer.getData(e.pixel);
 
         // TODO: Handle dataviews?
@@ -3856,22 +2097,22 @@ export class MainView extends React.Component<IMainViewProps, IStates> {
     const { layerId, layer: jgisLayer } = json;
     const isSourceType =
       typeof jgisLayer?.type === 'string' && jgisLayer.type.includes('Source');
-    const olLayer = this.getLayer(layerId);
+    const olLayer = this._mapViewer.getLayer(layerId);
 
     if (isSourceType) {
-      this.updateSource(layerId, jgisLayer);
+      this._mapViewer.updateSource(layerId, jgisLayer);
     }
     if (!jgisLayer || !olLayer) {
       this._log('error', 'Failed to update layer -- layer not found');
       return;
     }
-    this.updateLayer(layerId, jgisLayer, olLayer);
+    this._mapViewer.updateLayer(layerId, jgisLayer, olLayer);
   }
 
   private _convertFeatureToMs(_: IJupyterGISModel, args: string) {
     const json = JSON.parse(args);
     const { id: layerId, selectedFeature } = json;
-    const olLayer = this.getLayer(layerId);
+    const olLayer = this._mapViewer.getLayer(layerId);
     const source = olLayer.getSource() as VectorSource;
 
     if (typeof source.forEachFeature !== 'function') {
@@ -3892,7 +2133,7 @@ export class MainView extends React.Component<IMainViewProps, IStates> {
     const view = this._Map.getView();
     const zoom = view.getZoom();
     if (zoom) {
-      this._flyToPosition(newPosition, zoom);
+      this._mapViewer.flyToPosition(newPosition, zoom);
     } else {
       throw new Error(
         'Could not move to geolocation, because current zoom is not defined.',
@@ -4241,7 +2482,7 @@ export class MainView extends React.Component<IMainViewProps, IStates> {
     source: IJGISSource,
   ): Promise<void> => {
     this._model.sharedModel.updateSource(id, source);
-    await this.updateSource(id, source);
+    await this._mapViewer.updateSource(id, source);
   };
 
   private _renderAnnotationFloaters(): React.ReactNode {
@@ -4407,7 +2648,7 @@ export class MainView extends React.Component<IMainViewProps, IStates> {
   private storyViewerPanelRef = React.createRef<IStoryViewerPanelHandle>();
   private storyScrollContainerRef = React.createRef<HTMLDivElement>();
   private _Map: OlMap;
-  private _mapViewer: IMapViewer | null = null;
+  private _mapViewer: IMapViewer;
   private _lastMapViewerType: MapViewerType = 'openlayers';
   private _lastCenter: Coordinate | undefined;
   private _lastZoom: number | undefined;
@@ -4420,13 +2661,8 @@ export class MainView extends React.Component<IMainViewProps, IStates> {
   private _locationIndicatorActive = false;
   private _mainViewModel: MainViewModel;
   private _ready = false;
-  private _sources: Record<string, any>;
-  private _sourceToLayerMap = new Map();
   private _documentPath?: string;
   private _contextMenu: ContextMenu;
-  private _loadingLayers: Set<string>;
-  private _pendingZoomLayerId: string | null = null;
-  private _originalFeatures: IDict<Feature<Geometry>[]> = {};
   private _highlightLayerRef: {
     current: VectorImageLayer<VectorSource> | null;
   } = { current: null };
@@ -4446,8 +2682,9 @@ export class MainView extends React.Component<IMainViewProps, IStates> {
   private _annotationModel?: IAnnotationModel;
   private _loggerRegistry?: ILoggerRegistry;
   private _addLayerForPanels = (id: string, layer: IJGISLayer, index: number) =>
-    this.addLayer(id, layer, index);
-  private _removeLayerForPanels = (id: string) => this.removeLayer(id);
+    this._mapViewer.addLayer(id, layer, index);
+  private _removeLayerForPanels = (id: string) =>
+    this._mapViewer.removeLayer(id);
   private _patchGeoJSONFeatureAttributes: PatchGeoJSONFeatureAttributes;
 
   private _log(
