+++ packages/base/src/mainview/mapviewer.ts
@@ -1,9 +1,19 @@
-import type { Coordinate } from 'ol/coordinate';
+import type {
+  IDict,
+  IJGISLayer,
+  IJGISLayers,
+  IJGISSource,
+  IJGISSources,
+  IJupyterGISModel,
+  JgisCoordinates,
+} from '@jupytergis/schema';
 
 import { MapLibreViewer } from './viewers/maplibreViewer';
 import { OpenLayersViewer } from './viewers/openlayersViewer';
 
 export type MapViewerType = 'openlayers' | 'maplibre';
+
+export type MapCoordinate = [number, number];
 
 /**
  * Minimal abstraction layer between MainView and map engines.
@@ -15,22 +25,59 @@ export interface IMapViewer {
   initialize(target: HTMLElement, options: IMapViewerOptions): Promise<void>;
   destroy(): void;
 
-  getCenter(): Coordinate | undefined;
-  setCenter(center: Coordinate): void;
+  getCenter(): MapCoordinate | undefined;
+  setCenter(center: MapCoordinate): void;
   getZoom(): number | undefined;
   setZoom(zoom: number): void;
   getRotation(): number;
   setRotation(rotation: number): void;
   getProjection(): string;
 
+  getLayer(id: string): any | undefined;
+  getLayerIDs(): string[];
+  addLayer(id: string, layer: IJGISLayer, index?: number): Promise<void>;
+
+  removeLayer(id: string): void;
+
+  onZoomToPosition(_: IJupyterGISModel, id: string): void;
+  flyToPosition(
+    center: JgisCoordinates,
+    zoom: number,
+    duration?: number,
+    transitionType?: 'linear' | 'immediate' | 'smooth',
+  ): void;
+
+  updateLayersImpl(layerIds: string[]): Promise<void>;
+  trackLayerViewState(id: string, mapLayer: any): void;
+
+  updateLayer(
+    id: string,
+    layer: IJGISLayer,
+    mapLayer: any,
+    oldLayer?: IDict,
+  ): Promise<void>;
+  updateLayers(layerIds: string[]): void;
+
+  moveLayer(id: string, index: number): void;
+
+  addSource(id: string, source: IJGISSource): Promise<void>;
+
+  removeSource(id: string): void;
+
+  updateSource(id: string, source: IJGISSource): Promise<void>;
+
+  getSource(sourceId: string): IJGISSource | undefined;
+
   getViewport(): HTMLElement | undefined;
 }
 
 export interface IMapViewerOptions {
   projection?: string;
-  center?: Coordinate;
+  center?: MapCoordinate;
   zoom?: number;
   rotation?: number;
+  layers?: IJGISLayers;
+  sources?: IJGISSources;
 }
 
 /**
@@ -40,14 +87,15 @@ export interface IMapViewerOptions {
  */
 export async function createMapViewer(
   type: MapViewerType,
+  model: IJupyterGISModel,
 ): Promise<IMapViewer> {
   switch (type) {
     case 'openlayers': {
-      return new OpenLayersViewer();
+      return new OpenLayersViewer(model);
     }
 
     case 'maplibre': {
-      return new MapLibreViewer();
+      return new MapLibreViewer(model);
     }
 
     default: {
